добавлена возможность прикреплять к публикациям файлы, небольшие правки

+++ scientificWork/views.py
@@ -2,7 +2,8 @@
 # -*- coding: utf-8 -*-
 from django.contrib.auth import logout
 from django.core.exceptions import ValidationError
-from django.shortcuts import render
+from django.core.urlresolvers import reverse
+from django.shortcuts import render, render_to_response
 from django.contrib.auth import authenticate, login
 from django.http import HttpResponseRedirect, HttpResponse
 from django.template import loader, RequestContext
@@ -13,8 +14,8 @@ from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
 from report import *
-
-from scientificWork.models import Publication, UserProfile
+from django.core.files.storage import FileSystemStorage
+from scientificWork.models import Publication, UserProfile,MediaModel
 
 
 def index(request):
@@ -22,48 +23,53 @@ def index(request):
 
 
 def lk(request):
-    us = request.user
-    profile = UserProfile.objects.get(user=us)
-    profile_d = UserProfile_d(profile)
-    o = Publication.objects.all()
-    if 'deleteBook' in request.GET:
-        if o.filter(id=request.GET.get('deleteBook')).exists():
-            c = o.filter(id=request.GET.get('deleteBook'))
-            c.delete()
-    o = o.filter(user=profile)
-    o1 = []
-    for p in o:
-        o2 = Publication_d(p)
-        o1 += [o2]
-    template = loader.get_template('scientificWork/lk.html')
-    context = RequestContext(request, {
-        'profile': profile_d,
-        'publications': o1,
-    })
-    return HttpResponse(template.render(context))
+    if (request.user.is_authenticated()):
+        us = request.user
+        profile = UserProfile.objects.get(user=us)
+        profile_d = UserProfile_d(profile)
+        o = Publication.objects.all()
+        if 'deleteBook' in request.GET:
+            if o.filter(id=request.GET.get('deleteBook')).exists():
+                c = o.filter(id=request.GET.get('deleteBook'))
+                c.delete()
+        o = o.filter(user=profile)
+        o1 = []
+        for p in o:
+            o2 = Publication_d(p)
+            o1 += [o2]
+        template = loader.get_template('scientificWork/lk.html')
+        context = RequestContext(request, {
+            'profile': profile_d,
+            'publications': o1,
+        })
+        return HttpResponse(template.render(context))
+    else:
+        return HttpResponse("Пожалуйста авторизуйтесь")
 
 
 def addPublication(request):
-    '''o = Publication.objects.all()
-    c1 = Publication.objects.all()
-    c = c1.filter(bookName=u'блабла')
-    c.delete()'''
     if request.user.is_authenticated():
         st = 0
         if 'bookName' in request.GET:
             us = request.user
             profile = UserProfile.objects.get(user=us)
-            if len(request.GET.get('bookName')) > 0:
-                bookName = request.GET.get('bookName')
-                newp = Publication(user=profile, publishingHouseName="ETU", place="SPb",
-                                   date=datetime.datetime.strptime("2017-09-09", "%Y-%m-%d"), volume=3, unitVolume="p",
-                                   edition=9, type="book1", typePublication='book', bookName=bookName, isbn="jj",
-                                   number=1, editor="Andy", reiteration='disposable', citingBase='nil')
-                newp.save()
-                st = 1
+            bookName = request.GET.get('bookName')
+            typePublication=request.GET.get('type')
+            date=request.GET.get('date')
+            publishingHouseName=request.GET.get('publishingHouseName')
+            place=request.GET.get('place')
+            volume=request.GET.get('volume')
+            unitVolume=request.GET.get('unitVolume')
+            edition=request.GET.get('edition')
+            citingBase=request.GET.get('citingBase')
+            newp = Publication(user=profile, publishingHouseName=publishingHouseName, place=place,
+                                   date=datetime.datetime.strptime(date, "%Y-%m-%d"), volume=volume, unitVolume=unitVolume,
+                                   edition=edition, type="book1", typePublication=typePublication, bookName=bookName, isbn="jj",
+                                   number=1, editor="Andy", reiteration='disposable', citingBase=citingBase)
+            newp.save()
+            st = 1
 
-            else:
-                st = 2
+
         template = loader.get_template('scientificWork/addPublication.html')
         context = RequestContext(request, {
             'st': st,
@@ -78,8 +84,14 @@ def competitions(request):
 
 
 def publications(request):
+
     o = Publication.objects.all()
     filters = []
+    f = MediaModel.objects.all()
+    if 'deleteFile' in request.GET:
+        if f.filter(id=request.GET.get('deleteFile')).exists():
+            z=f.filter(id=request.GET.get('deleteFile'))
+            z.delete()
     if 'deleteBook' in request.GET:
         if o.filter(id=request.GET.get('deleteBook')).exists():
             c = o.filter(id=request.GET.get('deleteBook'))
@@ -226,4 +238,20 @@ def Publication_d(p):
     o2["author"] = p.user.patronymic
     o2["date"] = p.date
     o2["type"] = p.get_typePublication_display()
+    f=MediaModel.objects.all()
+    f1=f.filter(owner=p)
+    o2["f"]=f1
     return o2
+
+
+def upload_file(request):
+
+    if request.method == 'POST' and request.FILES['myfile']:
+        myfile = request.FILES['myfile']
+        M=MediaModel(media_file=myfile,owner=Publication.objects.get(id=request.GET.get("pub")))
+        M.save()
+        m = MediaModel.objects.all()
+        return render(request, 'scientificWork/upload_file.html', { 'pub': 0})
+    m = MediaModel.objects.all()
+    pub=request.GET.get("pub")
+    return render(request, 'scientificWork/upload_file.html', {'pubname':Publication.objects.get(id=request.GET.get("pub")).bookName, 'pub': pub})
